discloud/requirements.txt
discord.py
pyowm
